+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-21 15:53:37
-// Commit: style(state): harmonize light and dark theme palette for state
+// Updated at: 2026-07-21 17:53:51
+// Commit: feat(state): add bi-directional event bus handler for state
 
 class stateComponentState {
-  final String commitRef = '3002';
-  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
-  final DateTime timestamp = DateTime.parse('2026-07-21T15:53:37+05:00');
+  final String commitRef = '3007';
+  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
+  final DateTime timestamp = DateTime.parse('2026-07-21T17:53:51+05:00');
 }